fixing backend errors

+++ src/G74/DTO/CreateOperationRequestDTO.cs
@@ -1,13 +1,13 @@
 using G74.Domain.Value_Objects.Patient;
 
 public class CreateOperationRequestDTO
-{
+{   
     public string MedicalRecordNumber {get; set; }
     public string LicenceNumber {get; set; }
-    public int OperationTypeId { get; }
+    public long OperationTypeId { get; }
     public DateTime DeadlineDate {get;set; }
     public string Priority {get;set; }
-    public CreateOperationRequestDTO(string medicalRecordNumber, string licenceNumber, int operationTypeId, DateTime deadlineDate, string priority)
+    public CreateOperationRequestDTO(string medicalRecordNumber, string licenceNumber, long operationTypeId, DateTime deadlineDate, string priority)
     {
         MedicalRecordNumber = medicalRecordNumber;
         LicenceNumber = licenceNumber;
