+++ src/G74/DTO/OperationRequestDataModel.cs
@@ -4,9 +4,10 @@ namespace G74.DTO;
 
 public class OperationRequestDataModel : Entity<Guid>
 {
-    public int MedicalRecordNumber {get; set; }
-    public int LicenceNumber {get; set; }
-    public int OperationTypeId {get; set; }
+
+    public long MedicalRecordNumber {get; set; }
+    public long LicenceNumber {get; set; }
+    public long OperationTypeId {get; set; }
     public DateTime DeadlineDate {get; set; }
     public string Priority {get; set; }
 
@@ -16,8 +17,8 @@ public class OperationRequestDataModel : Entity<Guid>
     )
         : base(Guid.NewGuid()) 
     {
-        MedicalRecordNumber = int.Parse(request.MedicalRecordNumber.MedicalNumber);
-        LicenceNumber = int.Parse(request.LicenceNumber.licenceNumber);
+        MedicalRecordNumber = long.Parse(request.MedicalRecordNumber.MedicalNumber);
+        LicenceNumber = long.Parse(request.LicenceNumber.licenceNumber);
         OperationTypeId = request.OperationTypeId;
         DeadlineDate = request.DeadlineDate.date;
         Priority = request.Priority.PriorityDescription.ToString();
