+++ src/G74/Domain/Aggregates/OperationRequest/OperationRequest.cs
@@ -7,11 +7,11 @@ public class OperationRequest : IAggregateRoot
 {
     public MedicalRecordNumber MedicalRecordNumber {get; set; }
     public LicenceNumber LicenceNumber {get; set; }
-    public int OperationTypeId {get; set; }
+    public long OperationTypeId {get; set; }
     public DeadlineDate DeadlineDate {get; set; }
     public Priority Priority {get; set; }
 
-    public OperationRequest(MedicalRecordNumber medicalRecordNumber, LicenceNumber licenceNumber, int operationTypeId, DeadlineDate deadlineDate, Priority priority)
+    public OperationRequest(MedicalRecordNumber medicalRecordNumber, LicenceNumber licenceNumber, long operationTypeId, DeadlineDate deadlineDate, Priority priority)
     {
         MedicalRecordNumber = medicalRecordNumber;
         LicenceNumber = licenceNumber;
