+++ src/G74/Domain/Builders/OperationRequestBuilder.cs
@@ -5,14 +5,14 @@ public class OperationRequestBuilder
 {
     private MedicalRecordNumber _medicalRecordNumber ;
     private LicenceNumber _licenceNumber ;
-    private int _operationTypeId ;
+    private long _operationTypeId ;
     private DeadlineDate _deadlineDate ;
     private Priority _priority ;
 
     public OperationRequestBuilder(
         MedicalRecordNumber medicalRecordNumber,
         LicenceNumber licenceNumber,
-        int operationTypeId,
+        long operationTypeId,
         DeadlineDate deadlineDate,
         Priority priority
     ){
